Final Midwest code
works with ADC inputs, but the rainbow fader thing now seems to jump
between colors suddenly. Could this be due to the color brightness
adjustment?

Also, the ADC is struggling with data input since we have 100 ohm pull
downs with a 10k pot. We need to do some math or something to adjust for
the 100 ohm pulldowns that are messing with our data and making it
non-linear.

Also, we need a name for our box!